fix(basket): bounded baskets, ordered price sync and corrupt documents kept (Basket Stage 9, M2/M9/M10/L1/L6)
M2: a line holds at most 999 of a product and a basket at most 100 products
(D9), in both validators and in the domain. The merge on re-add is computed
in long, so a line stored before the limit cannot wrap around. TotalItems is
summed as long: two stored lines of 1.5 billion overflowed LINQ's checked
int Sum and that basket's GET failed for good. Stored documents are read as
they are; the limits apply to changes.

M10: PriceChangeWatermark records, per product, when the newest applied
price change happened (the event's OccurredOn). An older event arriving
later is skipped instead of putting the older price back; the second check,
after each basket read, makes an older event racing a newer one lose too.

M9: the price-sync fan-out re-prices up to 16 baskets at once instead of one
after another.

L1: ShoppingBasket.Create refuses an empty user id; Checkout sets
IsCheckedOut, so it cannot raise a second event and the basket takes no
further change; a line's identity is its product and its added time is
stored; LastModifiedAt reads the base class's UpdatedAt.

L6: a basket document that does not deserialize, belongs to another user
or holds a line the domain refuses still reads as no basket, but the write
that replaces or deletes it pushes it onto basket:corrupt:{userId} (10 kept,
30 days) in the same transaction and logs an error. A refused line used to
throw from every read.

+++ src/Services/Basket/EShop.Basket.Application/Queries/GetBasket/BasketDto.cs
@@ -9,7 +9,8 @@ public record BasketDto
     public string UserId { get; init; } = string.Empty;
     public List<BasketItemDto> Items { get; init; } = new();
     public decimal TotalPrice { get; init; }
-    public int TotalItems { get; init; }
+    /// <summary>A long: a basket stored before the per-line limit (Basket audit S9, M2) can exceed <c>int.MaxValue</c>.</summary>
+    public long TotalItems { get; init; }
 
     /// <summary>Null for a user who has no basket yet.</summary>
     public DateTime? CreatedAt { get; init; }
